+++ lib/tlv_decoder.dart
@@ -1,4 +1,4 @@
-library tlv_decoder;
+library;
 
 export 'src/model/tlv_model.dart';
 export 'src/tlv_utils.dart';